requirements.txt
numba
pillow
pygame
numpy